+++ q2_lefse/__init__.py
@@ -1,6 +1,6 @@
-from ._primary import run
-
-__version__ = "0.0.3"
+from ._primary import run, visualize
 from ._otu import OTUTable, OTUTableDirFmt, OTUTableFormat
 
-__all__ = ['run', 'OTUTable', 'OTUTableDirFmt', 'OTUTableFormat']
+__version__ = '0.1.3'
+
+__all__ = ['run', 'visualize', 'OTUTable', 'OTUTableDirFmt', 'OTUTableFormat']