+++ q2_lefse/plugin_setup.py
@@ -1,9 +1,7 @@
 import qiime2.plugin
-from q2_types.sample_data import SampleData
-from q2_types.per_sample_sequences import SequencesWithQuality
-from q2_types.feature_table import Frequency, RelativeFrequency, FeatureTable
 
 import q2_lefse
+from q2_lefse._otu import OTUTable, OTUTableDirFmt
 
 
 plugin = qiime2.plugin.Plugin(
@@ -11,30 +9,41 @@ plugin = qiime2.plugin.Plugin(
     version=q2_lefse.__version__,
     website='https://github.com/Science-notes/q2-lefse',
     package='q2_lefse',
-    user_support_text=("To get help with lefse, please post a question to "
-                       "https://github.com/Science-notes/q2-lefse"),
-    citation_text=None
+    user_support_text=(
+        'To get help with lefse, please post a question to '
+        'https://github.com/Science-notes/q2-lefse'
+    ),
+    citation_text=None,
 )
 
+plugin.register_semantic_types(OTUTable)
+plugin.register_formats(OTUTableDirFmt)
+plugin.register_semantic_type_to_format(OTUTable, artifact_format=OTUTableDirFmt)
+
+common_parameters = {
+    'class_id': qiime2.plugin.Int,
+    'subclass_id': qiime2.plugin.Int,
+    'subject_id': qiime2.plugin.Int,
+    'normalization': qiime2.plugin.Int,
+    'lda_threshold': qiime2.plugin.Float,
+    'wilcoxon_alpha': qiime2.plugin.Float,
+    'kruskal_alpha': qiime2.plugin.Float,
+    'command_prefix': qiime2.plugin.Str,
+}
 
 plugin.methods.register_function(
     function=q2_lefse.run,
-    inputs={'demultiplexed_seqs': SampleData[SequencesWithQuality]},
-    parameters={'threads': qiime2.plugin.Int},
-    name='Linear discriminant analysis Effect Size analysis using lefse',
-    outputs=[('genefamilies', FeatureTable[Frequency]),
-             ('pathcoverage', FeatureTable[RelativeFrequency]),
-             ('pathabundance', FeatureTable[RelativeFrequency])],
-    description='Execute the lefse'
+    inputs={'otu_table': OTUTable},
+    parameters=common_parameters,
+    name='Run LEfSe differential analysis',
+    outputs=[('lefse_results', OTUTable)],
+    description='Execute LEfSe (format + run) and produce the LEfSe result table.',
 )
 
 plugin.visualizers.register_function(
-    function=q2_lefse.run,
-    inputs={'demultiplexed_seqs': SampleData[SequencesWithQuality]},
-    parameters={'threads': qiime2.plugin.Int},
-    name='Linear discriminant analysis Effect Size analysis using lefse',
-    outputs=[('genefamilies', FeatureTable[Frequency]),
-             ('pathcoverage', FeatureTable[RelativeFrequency]),
-             ('pathabundance', FeatureTable[RelativeFrequency])],
-    description='Execute the lefse'
+    function=q2_lefse.visualize,
+    inputs={'otu_table': OTUTable},
+    parameters=common_parameters,
+    name='Visualize LEfSe output',
+    description='Run LEfSe and generate standard LEfSe plots and result table.',
 )